manage SDI node-selector of namespaces and dss
When dedicating nodes to SDI workload, one needs to annotate SDI
namespaces. On OCP cluster prior to 4.6, one also needs to set the same
node selector on daemonsets, otherwise the desired number of replicas
will match the number of schedulable nodes.

This is now automated with SDI Observer.

Added new template parameter SDI_NODE_SELECTOR. When unset, node
selectors of namespaces and daemonsets will not be managed. When set to
a comma-separated list of labels, the node selectors will be updated on
namespaces and daemonsets to match the value. When set to "removed",
the node selectors will be deleted.

Signed-off-by: Michal Minář <miminar@redhat.com>