observer: restore vsystem vrep
Let datahub operator manage the vsystem-vrep statefulset.
Otherwise, the operator may be unable to patch the statefulset because
e.g. volumeClaimTemplates cannot be updated in an existing resource.
This may prevent DI updates from succeeding.

The observer now patches datahub resource with
  vsystem.vRep.exportsMask = true
that causes the statefulset to use an emptyDir volume for /exports mount
point.

Signed-off-by: Michal Minář <miminar@redhat.com>